Refactore FileAppender to a package spark.util.logging and broke up the file into multiple files. Changed configuration parameter names.
